project_dashboard/requirements.txt
Django>=4.2.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
requests>=2.31.0
PyYAML>=6.0.1
python-decouple>=3.8
joblib>=1.3.0
PyMySQL>=1.1.0 
web3>=6.10.0